We have data!
It ain't pretty, but it works! Adds Tesla (like Guzzle, but for Elixir)
and the Hackney HTTP adapter to go with it (because apparently the
builtin one has issues verifying SSL certs or something), and uses that
to ping the Crunchyroll Queue API to get a list of the user's queue.

If this were Laravel, all that shit would be in the controller.